- Fixed can - not delete own previous leaves

+++ server/models/leaveModel.js
@@ -189,6 +189,9 @@ leaveSchema.statics.deleteLeave = async (leaveID, user) => {
     if (leaveRequest.user._id.toString() !== user._id.toString()) {
         throw new AppError("This leave is not belongs to your leave ,user can delete his own leave only", EC.ACCESS_DENIED, EC.HTTP_BAD_REQUEST)
     }
+    if (U.momentInUTC(leaveRequest.startDateString).isBefore(U.getNowMoment()) || U.momentInUTC(leaveRequest.startDateString).isBefore(U.getNowMoment())) {
+        throw new AppError("This leave already started or ended can not be delete", EC.ACCESS_DENIED, EC.HTTP_BAD_REQUEST)
+    }
 
     return await LeaveModel.findByIdAndRemove(mongoose.Types.ObjectId(leaveID))
 }
